Lemmatization, Progress Bars and Better User Feedback
Applied lemmatization with POS tagging to return words to their base form. It’s computationally heavy on weaker systems with larger datasets, so I added an input option for users to specify how many rows to preprocess. (Might add the option to choose a starting point—who knows?)
And for a little flair (and better user feedback, of course), I also added a progress bar for each preprocessing step